Fix to run past M = 0
Fix some mistakes that were giving wrong results. Generates some data to view intermediate results (IR), chooses default parameters in a way that makes null results (so we see that the parameterization was not set).

The main fix is to scalar density and kinectic energy so we can run past M = 0 and get results, not NaNs.