Map multiroot functions to proper range
The algorithm for multiroot finding assumes that the variables are
in the $(-\infty, +\infty)$ range, but this is not always true.
In this particular case, we have a variable in the $[0, 1]$ range
and another in the $[0, +\infty)$ range. This can be solved by
using a mapping:
	y = (tanh(x) + 1) / 2
	y = x^2

With that the algorithm avoids potentially problematic values
for the function we are trying to solve.

For m = 0, F_0 have a problem with the calculation of
$x^2\ln x$, resulting in NaN due to the logarithm. Solve
that by assign the proper limit (zero) to this term when
m = 0.