Some refactoring and options inclusion
The refactoring mainly reorganizes the rootfinding, so it is better
organized and reuses the one-dimensional rootfinder. A better
organization for the parameters is also used, now they are grouped.

Also adds option to change the proton fraction from commandline. The
parameters obtained from commandline are properly tested and set in a
function just for this purpose.

New rules are added to the makefile to facilitate running with beta
equilibrium and charge neutrality. Also silences the Makefile, so it don't keeps printing the recipes.

The README.md was updated to account for the new options and rules.